Prefixing jaas services
By prefxing jaas services we will be able identify them easily.
On the other hand, as we generate names instead of the daemon
we might end with collisions. In that case we will panic directly.

For the future we might want to pass names from commandline or retry in
case of a collision.

Signed-off-by: Dominik Dingel <dominik.dingel@gmail.com>